# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ounsa Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /me/settings:
    patch:
      operationId: updateSettings
      tags: [me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /me/password:
    patch:
      operationId: changePassword
      tags: [me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /me/welcome:
    post:
      operationId: setWelcomeReason
      tags: [me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WelcomeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /tasks/{key}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompleteResult"
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    SignupBody:
      type: object
      required: [name, password]
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        countryCode:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        password:
          type: string
        referralCode:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
    LoginBody:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, balance, totalInvested, totalEarned, totalWithdrawn, referralCode, language, hasDoubled, qualifiedReferralCount, wheelSpinsAvailable, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        countryCode:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        balance:
          type: string
        totalInvested:
          type: string
        totalEarned:
          type: string
        totalWithdrawn:
          type: string
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        hasDoubled:
          type: boolean
        qualifiedReferralCount:
          type: integer
        wheelSpinsAvailable:
          type: integer
        language:
          type: string
        welcomeReason:
          type: ["string", "null"]
        welcomeShownAt:
          type: ["string", "null"]
        createdAt:
          type: string
    UpdateSettingsBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
    WelcomeBody:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    ChangePasswordBody:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    TaskCompleteResult:
      type: object
      required: [reward, newBalance]
      properties:
        reward:
          type: string
        newBalance:
          type: string
...